LayoutTests/accessibility/aria-labelledby-overrides-aria-labeledby.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../fast/js/resources/js-test-style.css">
<script>
var successfullyParsed = false;
</script>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<button aria-labelledby="gamma" aria-labeledby="delta">Alpha</button>
<button aria-labeledby="epsilon">Beta</button>
<span id="gamma">Gamma</span>
<span id="delta">Delta</span>
<span id="epsilon">Epsilon</span>

<p id="description"></p>
<ul id="console"></ul>

<script>
    description("This tests that if aria-labelledby is used, then non-standard aria-labeledby attributes are ignored.");

    function log(str) {
        var console = document.getElementById("console");
        var li = document.createElement("li");
        li.appendChild(document.createTextNode(str));
        console.appendChild(li);
    }

    function logAXObject(obj) {
          log("for " + obj.allAttributes());
          log("children:\n" + obj.attributesOfChildren());
          log("obj.title is " + obj.title);
          log("obj.description is " + obj.description);
    }
    if (window.accessibilityController) {
          var body = document.getElementById("body");
          body.focus();
          var groupObj = accessibilityController.focusedElement.childAtIndex(0);
          var obj = groupObj.childAtIndex(0);
          // Title from aria-labelledby element, aria-labeledby is ignored
          shouldBe("obj.title", '"AXTitle: Gamma"');
          var obj = groupObj.childAtIndex(1);
          // aria-labeledby is used if no aria-labelledby present
          shouldBe("obj.title", '"AXTitle: Epsilon"');
    }

    successfullyParsed = true;
</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
